+++ src/binding/python/BaseRecord.cpp
@@ -26,24 +26,25 @@
 #include "openPMD/backend/MeshRecordComponent.hpp"
 #include "openPMD/backend/BaseRecordComponent.hpp"
 #include "openPMD/backend/PatchRecordComponent.hpp"
-
-#include <string>
+#include "openPMD/binding/python/UnitDimension.hpp"
 
 namespace py = pybind11;
 using namespace openPMD;
 
 
 void init_BaseRecord(py::module &m) {
-    py::class_<BaseRecord< RecordComponent >, Container< RecordComponent > >(m, "Base_Record_Record_Component")
-        .def_property_readonly("unit_dimension", &BaseRecord< RecordComponent >::unitDimension)
-        .def_property_readonly("scalar", &BaseRecord< RecordComponent >::scalar);
-    py::class_<BaseRecord< MeshRecordComponent >, Container< MeshRecordComponent > >(m, "Base_Record_Mesh_Record_Component")
-        .def_property_readonly("unit_dimension", &BaseRecord< MeshRecordComponent >::unitDimension)
-        .def_property_readonly("scalar", &BaseRecord< MeshRecordComponent >::scalar);
+    constexpr auto doc_scalar = R"docstr(
+Returns true if this record only contains a single component.
+)docstr";
+
     py::class_<BaseRecord< BaseRecordComponent >, Container< BaseRecordComponent > >(m, "Base_Record_Base_Record_Component")
-        .def_property_readonly("unit_dimension", &BaseRecord< BaseRecordComponent >::unitDimension)
-        .def_property_readonly("scalar", &BaseRecord< BaseRecordComponent >::scalar);
+        .def_property_readonly("unit_dimension", &BaseRecord< BaseRecordComponent >::unitDimension, python::doc_unit_dimension)
+        .def_property_readonly("scalar", &BaseRecord< BaseRecordComponent >::scalar, doc_scalar);
+
+    py::class_<BaseRecord< RecordComponent >, Container< RecordComponent > >(m, "Base_Record_Record_Component")
+        .def_property_readonly("scalar", &BaseRecord< RecordComponent >::scalar, doc_scalar);
+    py::class_<BaseRecord< MeshRecordComponent >, Container< MeshRecordComponent > >(m, "Base_Record_Mesh_Record_Component")
+        .def_property_readonly("scalar", &BaseRecord< MeshRecordComponent >::scalar, doc_scalar);
     py::class_<BaseRecord< PatchRecordComponent >, Container< PatchRecordComponent > >(m, "Base_Record_Patch_Record_Component")
-        .def_property_readonly("unit_dimension", &BaseRecord< PatchRecordComponent >::unitDimension)
-        .def_property_readonly("scalar", &BaseRecord< PatchRecordComponent >::scalar);
+        .def_property_readonly("scalar", &BaseRecord< PatchRecordComponent >::scalar, doc_scalar);
 }